Rename credential sanitizer module
